+++ src/treecf/__init__.py
@@ -28,7 +28,7 @@ from treecf.plausibility import Plausibility
 from treecf.regions import RecourseRegion
 from treecf.targets import Target
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 
 __all__ = [
     "AllowMissing",